+++ src/mru/store/model/BoardGames.java
@@ -1,17 +1,17 @@
 package mru.store.model;
 
 public class BoardGames extends Toys{
-	private int serialNum;
+	private long serialNum;
 	private String name;
 	private String brand;
-	private int price;
+	private float price;
 	private int stockCount;
 	private int minAge;
 	private int minNumOfPlayers;
 	private int maxNumOfPlayers;
 	private String designers; 
 	
-	public BoardGames(int serialNum, String name, String brand, int price, int stockCount, int minAge, int minNumOfPlayers, int maxNumOfPlayers, String designers) {
+	public BoardGames(long serialNum, String name, String brand, float price, int stockCount, int minAge, int minNumOfPlayers, int maxNumOfPlayers, String designers) {
 		this.serialNum = serialNum;
 		this.name = name;
 		this.brand = brand;
@@ -26,13 +26,13 @@ public class BoardGames extends Toys{
 	
 
 	@Override
-	public void setSerialNumber(int serialNum) {
+	public void setSerialNumber(long serialNum) {
 		this.serialNum = serialNum;
 	}
 
 
 	@Override
-	public int getSerialNumber() {
+	public long getSerialNumber() {
 		return serialNum;
 	
 	}
@@ -58,12 +58,12 @@ public class BoardGames extends Toys{
 	}
 
 	@Override
-	public void setPrice(int price) {
+	public void setPrice(float price) {
 		this.price = price;		
 	}
 
 	@Override
-	public int getPrice() {
+	public float getPrice() {
 		return price;
 	}
