+++ src/mru/store/model/Toys.java
@@ -5,8 +5,8 @@ public abstract class Toys {
 	//may need more
 	//may need to overload toString and other methods
 	
-	public abstract void setSerialNumber(int serialNum);
-	public abstract int getSerialNumber();
+	public abstract void setSerialNumber(long serialNum);
+	public abstract long getSerialNumber();
 
 	public abstract void setName(String name);
 	public abstract String getName();
@@ -14,8 +14,8 @@ public abstract class Toys {
 	public abstract void setBrand (String brand);
 	public abstract String getBrand ();
 
-	public abstract void setPrice(int price);
-	public abstract int getPrice();
+	public abstract void setPrice(float price);
+	public abstract float getPrice();
 
 	public abstract void setAvalibleCount(int stockCount);
 	public abstract int getAvalibleCount();
